chatbot/requirements.txt
streamlit
sql-agent-tool
sqlalchemy
psycopg2-binary
python-dotenv